style(audit): drop a redundant uint32 conversion
categoryFor already returns uint32; the conversion tripped unconvert.

+++ internal/audit/ocsf/class_per_record_test.go
@@ -62,7 +62,7 @@ func TestClassAndCategoryAgree(t *testing.T) {
 		class := classFor(a)
 		got := categoryFor(class)
 		want := class / 1000
-		if uint32(got) != want {
+		if got != want {
 			t.Errorf("%s: class %d sits in category %d; the class uid says %d",
 				a, class, got, want)
 		}
